reduce modes; change "button" to "switch"
first major galaxytopper software update:

I replaced the button in the band with a tilt-switch pointed nearly
vertically. Now when I tip the hat it changes patterns. I also set it up
to go back to stealth mode between patterns.

Patterns have been reduced to the following three:
 * Nebula  -- same as before
 * Rainbow -- same as before
 * Sparkle -- now randomly changes its speed and limits itself to highly
              saturated colors

Switched to a better HSV->RGB conversion

I've also got the beginnings of a "stars" mode that's supposed to
twinkle, but I need to do more work on that.